user can only delete his own cards

+++ controllers/cards.js
@@ -29,25 +29,25 @@ const createCard = (req, res) => {
     });
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(() => new Error('Not found'))
-    .then(() => {
-      res.send({ message: 'Card was successfully deleted' });
+const deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(() => {
+      throw new Error('Карточки с указанным _id не cуществует');
     })
-    .catch((err) => {
-      if (err.message === 'Not found') {
-        res.status(404).send({ message: 'Card not found' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Wrong ID format' });
-      } else {
-        res.status(500).send({
-          message: 'Could not delete card',
-          err: err.message,
-          stack: err.stack,
-        });
+    .then((card) => {
+      if (`${card.owner}` !== req.user._id) {
+        throw new Error('Удалять можно только свою карточку');
       }
-    });
+      Card.findByIdAndRemove(req.params.cardId)
+        .orFail(() => {
+          throw new Error('Карточка с указанным _id не найдена');
+        })
+        .then(() => {
+          res.send({ message: 'успешно' });
+        })
+        .catch((err) => next(err));
+    })
+    .catch((err) => next(err));
 };
 
 const likeCard = (req, res) => {
